+++ README.md
@@ -66,3 +66,13 @@ Additionally, make sure that the following extensions are enabled in your PHP:
 - json (enabled by default - don't turn it off)
 - [mysqlnd](http://php.net/manual/en/mysqlnd.install.php) if you plan to use MySQL
 - [libcurl](http://php.net/manual/en/curl.requirements.php) if you plan to use the HTTP\CURLRequest library
+
+## Marketplace Transactions
+
+Marketplace transaction features have been split into dedicated controllers to simplify maintenance:
+
+- `MarketplaceTransaction` – daftar, detail, dan statistik transaksi.
+- `MarketplaceTransactionImport` – mengelola proses impor Excel.
+- `MarketplaceTransactionTracking` – pelacakan dan pembaruan status resi.
+
+Lihat konfigurasi route di `app/Config/Routes.php` untuk detail endpoint terbaru.
